requirements.txt
kazoo==2.5.0
pykafka==2.8.0
python-dateutil==2.8.1
six==1.14.0
tabulate==0.8.7
text-unidecode==1.3
requests==2.31.0
pyyaml==6.0.1
pandas==1.5.3